+++ server/source/game/handlers/gm_items.rs
@@ -276,7 +276,7 @@ pub(super) async fn handle_slash_limpiar(state: &mut GameState, conn_id: Connect
                     tile.ground_item = world::GroundItem::default();
                 }
                 // Send BO (object delete) to everyone on that map
-                let bo_pkt = binary_packets::write_object_delete(x as u8, y as u8);
+                let bo_pkt = binary_packets::write_object_delete(x as i16, y as i16);
                 state.send_data_bytes(SendTarget::ToMap(map), &bo_pkt);
                 cleaned += 1;
             }
@@ -408,7 +408,7 @@ pub(super) async fn handle_slash_bloq(state: &mut GameState, conn_id: Connection
         let blocked_val = if is_blocked { 1 } else { 0 };
 
         // Broadcast BQ packet to everyone on the map
-        let bq_pkt = binary_packets::write_block_position(x as u8, y as u8, is_blocked);
+        let bq_pkt = binary_packets::write_block_position(x as i16, y as i16, is_blocked);
         state.send_data_bytes(SendTarget::ToMap(map), &bq_pkt);
 
         let status = if is_blocked { "bloqueado" } else { "desbloqueado" };
@@ -619,7 +619,7 @@ pub(super) async fn handle_slash_dest(state: &mut GameState, conn_id: Connection
     }
 
     // Send BO packet to notify clients
-    let bo_pkt = binary_packets::write_object_delete(x as u8, y as u8);
+    let bo_pkt = binary_packets::write_object_delete(x as i16, y as i16);
     state.send_data_bytes(SendTarget::ToMap(map), &bo_pkt);
 
     let gm_name = state.users.get(&conn_id).map(|u| u.char_name.clone()).unwrap_or_default();
@@ -659,7 +659,7 @@ pub(super) async fn handle_slash_massdest(state: &mut GameState, conn_id: Connec
         if let Some(tile) = grid.tile_mut(x, y) {
             tile.ground_item = world::GroundItem::default();
         }
-        let bo_pkt = binary_packets::write_object_delete(x as u8, y as u8);
+        let bo_pkt = binary_packets::write_object_delete(x as i16, y as i16);
         state.send_data_bytes(SendTarget::ToMap(map), &bo_pkt);
     }
 
@@ -719,7 +719,7 @@ pub(super) async fn handle_slash_haceritem(state: &mut GameState, conn_id: Conne
 
     // Send HO packet to show item visually
     if grh > 0 {
-        let ho_pkt = binary_packets::write_object_create(x as u8, y as u8, grh as i16);
+        let ho_pkt = binary_packets::write_object_create(x as i16, y as i16, grh as i16);
         state.send_data_bytes(SendTarget::ToArea { map, x, y }, &ho_pkt);
     }
