+++ server/source/game/handlers/skills/craft.rs
@@ -549,7 +549,7 @@ pub(crate) async fn handle_construct_smith(state: &mut GameState, conn_id: Conne
 
     // Play sound
     let (map, x, y) = state.users.get(&conn_id).map(|u| (u.pos_map, u.pos_x, u.pos_y)).unwrap_or((0,0,0));
-    let snd = binary_packets::write_play_wave(SND_HERRERO as u8, x as u8, y as u8);
+    let snd = binary_packets::write_play_wave(SND_HERRERO as u8, x as i16, y as i16);
     state.send_data_bytes(SendTarget::ToArea { map, x, y }, &snd);
 
     state.send_console(conn_id, &format!("Has construido {}", obj.name), font_index::INFO);
@@ -615,7 +615,7 @@ pub(crate) async fn handle_construct_carp(state: &mut GameState, conn_id: Connec
 
     // Play sound
     let (map, x, y) = state.users.get(&conn_id).map(|u| (u.pos_map, u.pos_x, u.pos_y)).unwrap_or((0,0,0));
-    let snd = binary_packets::write_play_wave(SND_CARPINTERO as u8, x as u8, y as u8);
+    let snd = binary_packets::write_play_wave(SND_CARPINTERO as u8, x as i16, y as i16);
     state.send_data_bytes(SendTarget::ToArea { map, x, y }, &snd);
 
     state.send_console(conn_id, &format!("Has construido {}", obj.name), font_index::INFO);
@@ -712,7 +712,7 @@ pub(crate) async fn handle_crear_fogata(state: &mut GameState, conn_id: Connecti
 
             // Get campfire GRH for visual
             let grh = state.get_object(FOGATA_OBJ).map(|o| o.grh_index).unwrap_or(0);
-            let ho_pkt = binary_packets::write_object_create(x as u8, y as u8, grh as i16);
+            let ho_pkt = binary_packets::write_object_create(x as i16, y as i16, grh as i16);
             state.send_data_bytes(SendTarget::ToArea { map, x, y }, &ho_pkt);
 
             // Add to cleanup list (temporary — VB6 uses garbage collector)
