+++ server/source/game/handlers/skills/resource.rs
@@ -52,7 +52,7 @@ pub(crate) async fn do_pescar(state: &mut GameState, conn_id: ConnectionId, tx: 
     }
 
     // Play sound to area
-    let snd = binary_packets::write_play_wave(SND_PESCAR as u8, ux as u8, uy as u8);
+    let snd = binary_packets::write_play_wave(SND_PESCAR as u8, ux as i16, uy as i16);
     state.send_data_bytes(SendTarget::ToArea { map, x: ux, y: uy }, &snd);
 
     // Luck roll
@@ -155,7 +155,7 @@ pub(crate) async fn do_talar(state: &mut GameState, conn_id: ConnectionId, tx: i
     }
 
     // Play sound
-    let snd = binary_packets::write_play_wave(SND_TALAR as u8, ux as u8, uy as u8);
+    let snd = binary_packets::write_play_wave(SND_TALAR as u8, ux as i16, uy as i16);
     state.send_data_bytes(SendTarget::ToArea { map, x: ux, y: uy }, &snd);
 
     // Luck roll
@@ -257,7 +257,7 @@ pub(crate) async fn do_mineria(state: &mut GameState, conn_id: ConnectionId, tx:
     }
 
     // Play sound
-    let snd = binary_packets::write_play_wave(SND_MINERO as u8, ux as u8, uy as u8);
+    let snd = binary_packets::write_play_wave(SND_MINERO as u8, ux as i16, uy as i16);
     state.send_data_bytes(SendTarget::ToArea { map, x: ux, y: uy }, &snd);
 
     // Luck roll
